Split out omega block pasing into a separate header file
